+++ noodles-bcf/src/async/reader/record.rs
@@ -66,7 +66,7 @@ mod tests {
         read_record(&mut reader, &mut buf, &mut record).await?;
 
         assert_eq!(record.chromosome_id(), 1);
-        assert_eq!(record.position(), Position::try_from(101)?);
+        assert_eq!(record.position(), Position::from(101));
         assert_eq!(record.rlen(), 1);
         assert_eq!(
             record.quality_score(),