+++ noodles-bcf/src/record/convert.rs
@@ -23,7 +23,7 @@ impl Record {
     /// let actual = record.try_into_vcf_record(&header, &string_maps)?;
     /// let expected = vcf::Record::builder()
     ///     .set_chromosome("sq0".parse()?)
-    ///     .set_position(Position::try_from(1)?)
+    ///     .set_position(Position::from(1))
     ///     .set_reference_bases("A".parse()?)
     ///     .build()?;
     ///