+++ noodles-bcf/src/writer.rs
@@ -135,7 +135,7 @@ where
     ///
     /// let record = vcf::Record::builder()
     ///     .set_chromosome("sq0".parse()?)
-    ///     .set_position(Position::try_from(8)?)
+    ///     .set_position(Position::from(8))
     ///     .set_reference_bases("A".parse()?)
     ///     .build()?;
     ///